Refactor profile form store import paths
