+++ app/onboarding/profile/index.tsx
@@ -12,7 +12,7 @@ import { Label } from "~/components/ui/label";
 import { Text } from "~/components/ui/text";
 import { formatDate } from "~/lib/date";
 import { fontFamily } from "~/lib/font";
-import { useProfileFormStore } from "~/app/profile/edit/profile-form-store";
+import { useProfileFormStore } from "~/stores/profile-form-store";
 import { useShallow } from "zustand/react/shallow";
 import { Button } from "~/components/ui/button";
 import { useMutation } from "convex/react";
